Using same width and height for viewport and window

+++ src/dust/camera.rs
@@ -12,24 +12,24 @@ pub struct Camera {
     direction: glm::Vec3,
     z_near: f32,
     z_far: f32,
-    width: i32,
-    height: i32,
+    width: u32,
+    height: u32,
 }
 
 
 impl Camera
 {
-    pub fn create(gl: &gl::Gl, position: glm::Vec3, direction: glm::Vec3, width: i32, height: i32) -> Result<Camera, Error>
+    pub fn create(gl: &gl::Gl, position: glm::Vec3, direction: glm::Vec3, width: u32, height: u32) -> Result<Camera, Error>
     {
         let mut camera = Camera { gl: gl.clone(), position: position, direction: direction, z_near: 0.1, z_far: 1000.0, width: width, height: height };
         camera.set_screen_size(width, height);
         Ok(camera)
     }
 
-    pub fn set_screen_size(&mut self, width: i32, height: i32)
+    pub fn set_screen_size(&mut self, width: u32, height: u32)
     {
         unsafe {
-            self.gl.Viewport(0, 0, width, height);
+            self.gl.Viewport(0, 0, width as i32, height as i32);
         }
         self.width = width;
         self.height = height;
